Unit test RuntimeMetadataService
