+++ src/services/accounts/AccountsBalanceInfoService.spec.ts
@@ -1,5 +1,5 @@
 import { sanitizeNumbers } from '../../sanitize/sanitizeNumbers';
-import { blockHash789629, mockApi, testAddress } from '../mock/mockApi';
+import { blockHash789629, mockApi, testAddress } from '../mock';
 import * as accountsBalanceInfo789629 from './accountsBalanceInfo789629.json';
 import { AccountsBalanceInfoService } from './AccountsBalanceInfoService';
