+++ src/services/mock/testAddress.ts
@@ -1,0 +1,4 @@
+/**
+ * Address to use with Accounts tests.
+ */
+export const testAddress = `1zugcapKRuHy2C1PceJxTvXWiq6FHEDm2xa5XSU7KYP3rJE`;
